requirements.txt
python-dotenv
requests
Flask
flask-login
psycopg2-binary
Flask-sqlalchemy==2.1